+++ src/Storage/Storage.Management/File/NewAzureStorageShare.cs
@@ -87,6 +87,29 @@ namespace Microsoft.Azure.Commands.Management.Storage
         [ValidateNotNull]
         public Hashtable Metadata { get; set; }
 
+
+        [Parameter(
+           Mandatory = false,
+           HelpMessage = "Access tier for specific share. StorageV2 account can choose between TransactionOptimized (default), Hot, and Cool. FileStorage account can choose Premium.")]
+        [ValidateSet(ShareAccessTier.TransactionOptimized,
+            ShareAccessTier.Premium,
+            ShareAccessTier.Hot,
+            ShareAccessTier.Cool,
+           IgnoreCase = true)]
+        [ValidateNotNullOrEmpty]
+        public string AccessTier
+        {
+            get
+            {
+                return accessTier;
+            }
+            set
+            {
+                accessTier = value;
+            }
+        }
+        private string accessTier = null;
+        
         public override void ExecuteCmdlet()
         {
             base.ExecuteCmdlet();
@@ -113,7 +136,8 @@ namespace Microsoft.Azure.Commands.Management.Storage
                             this.Name,
                             new FileShare(
                                 metadata: MetadataDictionary,
-                                shareQuota: shareQuota));
+                                shareQuota: shareQuota,
+                                accessTier: accessTier));
 
                 WriteObject(new PSShare(share));
             }